// src/app/models/parsed-detail.model.ts
export class ParsedDetailModel {
  name: string;
  abilities: string;
  baseExperience: number;
  height: number;
  heldItems: string;
  moves: string;
  species: string;
  weight: number;
  types: string;
  hp: any;
  speed: any;
  attack: any;
  defense: any;
  frontImage: any;
  backImage: any;
  frontShinyImage: any;
  backShinyImage: any;
  frontImageF: any;
  backImageF: any;
  frontShinyImageF: any;
  backShinyImageF: any;
}